Extract common docs tooling
