+++ enterprise/ha/src/test/java/jmx/JmxDocTest.java
@@ -48,13 +48,13 @@ import javax.management.ObjectInstance;
 import javax.management.ObjectName;
 
 import org.neo4j.cluster.ClusterSettings;
+import org.neo4j.doc.tools.AsciiDocGenerator;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.factory.GraphDatabaseBuilder;
 import org.neo4j.graphdb.factory.TestHighlyAvailableGraphDatabaseFactory;
 import org.neo4j.helpers.collection.Iterators;
 import org.neo4j.kernel.configuration.docs.SettingDescription;
 import org.neo4j.kernel.configuration.docs.AsciiDocListGenerator;
-import org.neo4j.test.AsciiDocGenerator;
 import org.neo4j.test.TargetDirectory;
 
 import static org.junit.Assert.assertEquals;
