+++ manual/cypher/cypher-docs/pom.xml
@@ -149,6 +149,14 @@
       <scope>test</scope>
     </dependency>
 
+    <dependency>
+      <groupId>org.neo4j.doc</groupId>
+      <artifactId>neo4j-doc-tools</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+      <type>test-jar</type>
+    </dependency>
+
   </dependencies>
 
     <build>
