+++ manual/embedded-examples/pom.xml
@@ -96,6 +96,14 @@
       <version>${project.version}</version>
     </dependency>
 
+    <dependency>
+      <groupId>org.neo4j.doc</groupId>
+      <artifactId>neo4j-doc-tools</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+      <type>test-jar</type>
+    </dependency>
+
   </dependencies>
 
   <licenses>
