+++ manual/embedded-examples/src/test/java/org/neo4j/examples/AbstractJavaDocTestBase.java
@@ -25,11 +25,11 @@ import org.junit.Rule;
 
 import java.util.Map;
 
+import org.neo4j.doc.tools.JavaTestDocsGenerator;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Node;
 import org.neo4j.test.GraphDescription;
 import org.neo4j.test.GraphHolder;
-import org.neo4j.test.JavaTestDocsGenerator;
 import org.neo4j.test.TestData;
 import org.neo4j.visualization.asciidoc.AsciidocHelper;
