+++ manual/server-examples/pom.xml
@@ -108,6 +108,15 @@
       <artifactId>jackson-mapper-asl</artifactId>
       <version>1.9.7</version>
     </dependency>
+
+    <dependency>
+      <groupId>org.neo4j.doc</groupId>
+      <artifactId>neo4j-doc-tools</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+      <type>test-jar</type>
+    </dependency>
+
   </dependencies>
 
   <licenses>
