+++ pom.xml
@@ -42,6 +42,7 @@
     <module>integrationtests</module>
     <module>stresstests</module>
     <module>tools</module>
+    <module>doctools</module>
   </modules>
 
   <build>
